Move App into its own file and out of main
