+++ homelab_backend/src/api/games_controller.rs
@@ -2,7 +2,7 @@ use std::sync::Arc;
 
 use super::get_user_from_auth_header;
 use super::return_data::ReturnData;
-use crate::AppState;
+use crate::app::AppState;
 use axum::{
     extract::{Path, State},
     http::header::HeaderMap,
