+++ homelab_backend/src/api/reminder_controller.rs
@@ -1,6 +1,6 @@
 use super::get_user_from_auth_header;
 use super::return_data::ReturnData;
-use crate::AppState;
+use crate::app::AppState;
 use axum::{
     extract::{Path, State},
     http::header::HeaderMap,
