+++ homelab_backend/src/api/user_controller.rs
@@ -11,13 +11,13 @@ use sha2::{Digest, Sha256};
 
 use crate::{
     api::{get_user_from_auth_header, return_data::ReturnData},
+    app::AppState,
     models::user::{
         jwt::encode_jwt,
         user_db::{db_create_user, db_delete_user, db_get_user_by_id, db_get_user_by_username, db_update_user},
         validation::{LoginUserSchema, UpdateUserSchema},
         AuthLevel, ReturnUser,
     },
-    AppState,
 };
 
 const SALT_LENGTH: usize = 12;
